Added new classes, tested changes in main method
I added the ExitView, Win, and Lose classes, and I added a few lines to the main method in the Game class to test the new things I added. I added constructors to Win and Lose classes, and changed the access modifiers of some methods to try and allow our classes to communicate with each other and utilize the methods within. Game class creates an object of type Win, and this object is then able to access the methods within the Win class to check if the correct word has been guessed, and call the ExitView if the game is over. Lose class functions similarly, but will need additional functionality to track how many rounds have passed. ExitView class is not currently functional. Filled out some of the methods in the game class to enable communication between Game and Win, as well as Game and Lose.