Add Game View File
I put together a prototype for how I think the game could look. It still needs the logic for displaying words as the game is played, checking the letters within the words for correctness, and also the code for allowing this class to communicate with the Game class. I hope that this helps you though.